README.md
I have created as the first project of my Boothcamp. It's a fusion of a game and a movie, both from my childhood. They are the game of Space invaders and Star Wars. 

I decided to create a story which begins to appear from the first battle "the Impire Strike Back".

For this project I code mostly in JavaScript, HTML, CSS.